+++ src/apps/idle/ethereum/idle.vault.token-fetcher.ts
@@ -1,98 +1,56 @@
 import { Inject } from '@nestjs/common';
-import _ from 'lodash';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
-import { Register } from '~app-toolkit/decorators';
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+import { AppTokenTemplatePositionFetcher } from '~position/template/app-token.template.position-fetcher';
 import {
-  buildDollarDisplayItem,
-  buildPercentageDisplayItem,
-} from '~app-toolkit/helpers/presentation/display-item.present';
-import { getAppImg } from '~app-toolkit/helpers/presentation/image.present';
-import { ContractType } from '~position/contract.interface';
-import { PositionFetcher } from '~position/position-fetcher.interface';
-import { AppTokenPosition } from '~position/position.interface';
-import { Network } from '~types/network.interface';
+  GetDataPropsParams,
+  GetPricePerShareParams,
+  GetUnderlyingTokensParams,
+} from '~position/template/app-token.template.types';
 
-import { IdleContractFactory } from '../contracts';
-import { IDLE_DEFINITION } from '../idle.definition';
+import { IdleContractFactory, IdleToken } from '../contracts';
 
-const appId = IDLE_DEFINITION.id;
-const groupId = IDLE_DEFINITION.groups.vault.id;
-const network = Network.ETHEREUM_MAINNET;
+@PositionTemplate()
+export class EthereumIdleVaultTokenFetcher extends AppTokenTemplatePositionFetcher<IdleToken> {
+  groupLabel = 'Vault';
+  isExcludedFromBalances = true;
+  isExcludedFromExplore = true;
+  isExcludedFromTvl = true;
 
-@Register.TokenPositionFetcher({ appId, groupId, network })
-export class EthereumIdleVaultTokenFetcher implements PositionFetcher<AppTokenPosition> {
   constructor(
-    @Inject(IdleContractFactory) private readonly idleContractFactory: IdleContractFactory,
-    @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
-  ) {}
+    @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
+    @Inject(IdleContractFactory) protected readonly contractFactory: IdleContractFactory,
+  ) {
+    super(appToolkit);
+  }
 
-  async getPositions() {
-    const multicall = this.appToolkit.getMulticall(network);
-    const baseTokens = await this.appToolkit.getBaseTokenPrices(network);
+  getContract(address: string): IdleToken {
+    return this.contractFactory.idleToken({ network: this.network, address });
+  }
 
-    const controller = this.idleContractFactory.idleController({
+  async getAddresses() {
+    const controller = this.contractFactory.idleController({
       address: '0x275da8e61ea8e02d51edd8d0dc5c0e62b4cdb0be',
-      network,
+      network: this.network,
     });
-    const marketTokenAddresses = await multicall.wrap(controller).getAllMarkets();
 
-    const vaultTokens = await Promise.all(
-      marketTokenAddresses.map(async tokenAddressRaw => {
-        const tokenAddress = tokenAddressRaw.toLowerCase();
-        const tokenContract = this.idleContractFactory.idleToken({ network, address: tokenAddress });
+    return controller.getAllMarkets();
+  }
 
-        const [label, symbol, underlyingTokenAddressRaw, decimalsRaw, priceRaw, apyRaw, supplyRaw] = await Promise.all([
-          multicall.wrap(tokenContract).name(),
-          multicall.wrap(tokenContract).symbol(),
-          multicall.wrap(tokenContract).token(),
-          multicall.wrap(tokenContract).decimals(),
-          multicall.wrap(tokenContract).tokenPrice(),
-          multicall.wrap(tokenContract).getAvgAPR(),
-          multicall.wrap(tokenContract).totalSupply(),
-        ]);
+  async getUnderlyingTokenAddresses({ contract }: GetUnderlyingTokensParams<IdleToken>) {
+    return contract.token();
+  }
 
-        const underlyingToken = baseTokens.find(price => price.address === underlyingTokenAddressRaw.toLowerCase());
-        if (!underlyingToken) return null;
+  async getPricePerShare({ appToken, contract }: GetPricePerShareParams<IdleToken>) {
+    const priceRaw = await contract.tokenPrice();
+    const price = Number(priceRaw) / 10 ** appToken.tokens[0].decimals;
 
-        const decimals = Number(decimalsRaw);
-        const price = Number(priceRaw) / 10 ** underlyingToken.decimals;
-        const pricePerShare = price / underlyingToken.price;
-        const apy = Number(apyRaw) / 10 ** 18 / 100;
-        const supply = Number(supplyRaw) / 10 ** decimals;
-        const liquidity = supply * price;
-        const tokens = [underlyingToken];
+    return price / appToken.tokens[0].price;
+  }
 
-        const appToken: AppTokenPosition = {
-          type: ContractType.APP_TOKEN,
-          network,
-          appId,
-          groupId,
-          address: tokenAddress,
-          decimals,
-          symbol,
-          displayProps: {
-            label,
-            statsItems: [
-              { label: 'Liquidity', value: buildDollarDisplayItem(liquidity) },
-              { label: 'APY', value: buildPercentageDisplayItem(apy) },
-            ],
-            images: [getAppImg(appId)],
-          },
-          dataProps: {
-            liquidity,
-            apy,
-          },
-          price,
-          pricePerShare,
-          supply,
-          tokens,
-        };
-
-        return appToken;
-      }),
-    );
-
-    return _.compact(vaultTokens);
+  async getApy({ contract }: GetDataPropsParams<IdleToken>): Promise<number> {
+    const apyRaw = await contract.getAvgAPR();
+    return Number(apyRaw) / 10 ** 18 / 100;
   }
 }